requirements.txt
beautifulsoup4
lxml
requests-html
python-telegram-bot